refactor cage generation codes to different class files
